Update ItemDropSpawner and Add Item Exp

+++ Assets/Script/Item/DropRate.cs
@@ -1,13 +1,11 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 using System;
+using UnityEngine;
 
 [Serializable]
 public class DropRate
 {
     public ItemSO itemSO;
-    public float dropRate;
+    public int dropRate;
     public int minDrop;
     public int maxDrop;
 }
